# swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/admin/client":{"put":{"tags":["admin-controller"],"operationId":"editClientByAddress","parameters":[{"name":"address","in":"query","required":true,"schema":{"pattern":"^https?://(www\\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_+.~#?&/=]*)$","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["admin-controller"],"operationId":"addClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"delete":{"tags":["admin-controller"],"operationId":"deleteClientByAddress","parameters":[{"name":"address","in":"query","required":true,"schema":{"pattern":"^https?://(www\\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_+.~#?&/=]*)$","type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}}}},"/api/v1/admin/provider":{"patch":{"tags":["admin-controller"],"operationId":"setStatus","parameters":[{"name":"status","in":"query","required":true,"schema":{"type":"boolean"}},{"name":"providerCredential","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/admin/provider/trust-coefficient":{"patch":{"tags":["admin-controller"],"operationId":"setTrustCoefficient","parameters":[{"name":"trustCoefficient","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"providerCredential","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/report/weekly/{providerCredential}":{"get":{"tags":["report-controller"],"operationId":"getWeeklyReport","parameters":[{"name":"providerCredential","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/report/monthly/{providerCredential}":{"get":{"tags":["report-controller"],"operationId":"getMonthlyReport","parameters":[{"name":"providerCredential","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/provider":{"get":{"tags":["provider-controller"],"operationId":"getProviders","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProviderDTO"}}}}}}}},"/api/v1/provider/active":{"get":{"tags":["provider-controller"],"operationId":"getActiveProviders","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProviderDTO"}}}}}}}},"/api/v1/forecast/type":{"get":{"tags":["forecast-controller"],"operationId":"getForecastTypes","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/api/v1/forecast/provider/{providerCredential}":{"get":{"tags":["forecast-controller"],"operationId":"getCityForecastFromProvider","parameters":[{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/ForecastRequest"}},{"name":"providerCredential","in":"path","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"maximum":100,"type":"integer","format":"int32","default":50}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ForecastPageDTO"}}}}}}},"/api/v1/forecast/measurement":{"get":{"tags":["forecast-controller"],"operationId":"getMeasurementUnits","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string","enum":["METRIC","US","UK"]}}}}}}}},"/api/v1/forecast/aggregated":{"get":{"tags":["forecast-controller"],"operationId":"getCityForecastAggregated","parameters":[{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/ForecastRequest"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"maximum":100,"type":"integer","format":"int32","default":50}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ForecastPageDTO"}}}}}}},"/api/v1/city/name":{"get":{"tags":["city-controller"],"operationId":"getCitiesByName","parameters":[{"name":"cityName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CityDTO"}}}}}}}},"/api/v1/city/coordinates":{"get":{"tags":["city-controller"],"operationId":"getCityByCoordinates","parameters":[{"name":"lat","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"lng","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExceptionResponseDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CityDTO"}}}}}}}},"components":{"schemas":{"ExceptionResponseDTO":{"type":"object","properties":{"exceptionMessage":{"type":"string"},"requestDescription":{"type":"string"},"httpStatus":{"type":"integer","format":"int32"},"date":{"type":"string"}}},"ClientDTO":{"required":["address","name"],"type":"object","properties":{"address":{"pattern":"^https?://(www\\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_+.~#?&/=]*)$","type":"string"},"name":{"type":"string"}}},"SignUpDTO":{"required":["address","name","password"],"type":"object","properties":{"name":{"maxLength":64,"minLength":4,"type":"string"},"password":{"maxLength":64,"minLength":10,"pattern":"^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[a-zA-Z]).{8,}$","type":"string"},"address":{"pattern":"^https?://(www\\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_+.~#?&/=]*)$","type":"string"}}},"TokenDTO":{"required":["accessToken","refreshToken"],"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}},"LoginDTO":{"required":["password","username"],"type":"object","properties":{"username":{"maxLength":64,"minLength":4,"type":"string"},"password":{"maxLength":64,"minLength":10,"type":"string"}}},"ProviderDTO":{"type":"object","properties":{"credential":{"type":"string"},"active":{"type":"boolean"}}},"ForecastRequest":{"required":["endDate","startDate","type"],"type":"object","properties":{"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"type":{"type":"string"},"unit":{"type":"string","enum":["METRIC","US","UK"]},"lat":{"type":"number","format":"double"},"lng":{"type":"number","format":"double"}}},"CityDTO":{"type":"object","properties":{"name":{"type":"string"},"lat":{"type":"number","format":"double"},"lng":{"type":"number","format":"double"},"country":{"type":"string"}}},"ForecastDTO":{"type":"object","properties":{"windSpeed":{"type":"number","format":"double"},"windDirection":{"type":"number","format":"double"},"temperature":{"type":"number","format":"double"},"pressure":{"type":"number","format":"double"},"humidity":{"type":"number","format":"double"},"time":{"type":"string","format":"date-time"},"conditions":{"type":"string"}}},"ForecastPageDTO":{"type":"object","properties":{"forecastDTOs":{"$ref":"#/components/schemas/PageForecastDTO"},"timeStep":{"type":"string"},"cityDTO":{"$ref":"#/components/schemas/CityDTO"},"providerCredential":{"type":"string"}}},"PageForecastDTO":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ForecastDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}}}}}